requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
elasticsearch==5.1.0
elasticsearch-dsl==5.1.0
fake-useragent==0.1.11
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kiwisolver==1.0.1
lxml==4.3.0
matplotlib==3.0.2
mysqlclient==1.3.14
numpy==1.16.0
parsel==1.5.1
parso==0.3.1
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.3.1
pypiwin32==223
python-dateutil==2.7.5
pywin32==224
queuelib==1.5.0
redis==3.2.0
requests==2.21.0
Scrapy==1.5.1
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
wcwidth==0.1.7
zope.interface==4.6.0
